Provide configurability in the location and naming of the uploaded artifacts
[Problem]
To specify the code S3 location for a lambda function, the location must be deterministic. The locations of the
OutputArtifacts of CodeBuild is not.

[Solution]
Use environment variables specified in the CloudFormation CodeBuild project to specify where the artifacts are to be
uploaded to.

[Test]
Updated the stack with it pointing at a different `S3DropboxLambda` repository branch - the code was uploaded to the
expected S3 location.