Publish zip'd S3 files to Dropbox
[Problem]
The files uploaded to S3, originating from the output of the dossiers repository, are not uploaded to Dropbox.

[Solution]
Upload the files from S3 to Dropbox.

As part of this change, the CloudFormation architecture was changed somewhat, which had the following impact:
  * The function configuration is updated as part of the CodeBuild build stage
  * The Lambda function is updated via the `aws lambda update-function-code` instead of simply uploading to S3. This,
      technically, was a bug because Lambda code is not updated simply by uploading a new package to the S3 location.

[Test]
Used the code in the new CloudFormation pipeline. The files were successfully published to Dropbox.