Only update files if they have been modified later than that found on Dropbox
[Problem]
All of the files in the Dropbox folder are updated when a change is made to the package. Instead, only the updated PDFs
should be updated.

[Solution]
The latex build has been changed so that the PDF has the same timestamp of the LaTeX file. Thus, by adding this
timestamp to the update, Dropbox should ignore the file if it has not been changed.

[Test]
* added timestamp validation to the ZipFileIteratorSpec