Use manifest file to determine whether update is required
[Problem]
The timestamps of files are lost during the CodeCommit packaging of the source code. This requires a different paradigm
to determine whether a file is needed to be updated.

[Solution]
Keep a manifest file in the Dropbox App folder, which contains a hash of the LaTeX file. This hash is compared against
the hash of the new LaTeX file in the payload. If they are not the same, then it can be assumed that the PDF needs to be
updated. Note that the the PDF can be updated with the date during the build, making it unusable as the hash source.

[Test]
* Add unit tests covering the new files
* Manually uploaded and tested: uploaded same zip file and no updates were made, neither to the PDFs nor the manifest.
    This was determined by looking at the Lambda function's CloudWatch logs.