Use short lived Dropbox tokens
## Problem
Dropbox is deprecating long lived access tokens.

* https://dropbox.tech/developers/migrating-app-permissions-and-access-tokens

## Solution
Migrate application to use short lived applications by storing the credentials in S3 encrypted with KMS. Tokens are
refreshed at the beginning of the application and updated in S3 immediately for the next refresh, which will occur when
the application is run again.

A number of improvements were made during development:

1. Spring-ified application for dependency injection
2. Added support for running application locally for testing
3. Increased test coverage