Retrieve manifest from build and support folder hierarchy
## Problem
With the increase in LaTeX documents in different subject areas, the flattened document list in Dropbox is becoming
unwieldy and leaves a desire for a hierarchical organization of these documents.

## Solution
The LaTeX build was updated to publish hierarchical PDF documents and publishes the manifest to go along with these
PDFs. The manifest is calculated using the same logic as was found here i.e. taking the MD5 sum of all the source '.tex'
documents.

This application was updated to use the source manifest instead of calculating its own and updated to support documents
in a folder structure. When documents are deleted, it recursively deletes folders when finding that they are empty.

## Test
* unit tests
* manually tested via CLI main