Add multi-threading to Dropbox file upload
## Problem
The Lambda function times out the first time that the documents are uploaded to Dropbox.

## Solution
Looking at the AWS Lambda logs shows that the unzip'ing of LaTeX package takes about 7 seconds. Uploading the files to
Dropbox, which was sequential was taking 4.5 minutes and causing the Lambda to timeout since the timeout was set to 5
minutes.

This change adds parallelism in uploading the files to Dropbox since we are likely not saturating the network bandwidth.

## Test
* unit tests
* ran locally with sbt